requirements_ai.txt
# AI功能依赖包
ultralytics>=8.0.0  # YOLOv8支持
torch>=1.9.0        # PyTorch深度学习框架
torchvision>=0.10.0 # PyTorch视觉库
opencv-python>=4.5.0 # OpenCV图像处理
numpy>=1.21.0       # 数值计算
pillow>=8.0.0       # 图像处理
matplotlib>=3.3.0   # 数据可视化
pyyaml>=5.4.0       # YAML配置文件
psutil>=5.8.0       # 系统监控(可选)

# 现有依赖包
pyqt5>=5.14.1       # GUI框架
lxml>=4.9.1         # XML处理